requirements.txt
<<<<<<< HEAD
flask
flask-cors
python-dotenv
numpy==1.26.4
scikit-learn==1.3.0
pandas==2.2.2
joblib==1.4.2
opencv-python==4.9.0.80
fer==22.5.1
speechrecognition==3.10.1
pyaudio==0.2.14
pillow==10.4.0
requests
scipy==1.11.4
moviepy==1.0.3
tensorflow-cpu==2.15.0
psycopg2-binary
bcrypt
nltk
SQLAlchemy
=======
flask
flask-cors
python-dotenv
numpy==1.26.4
scikit-learn==1.3.0
pandas==2.2.2
joblib==1.4.2
opencv-python==4.9.0.80
fer==22.5.1
speechrecognition==3.10.1
pyaudio==0.2.14
pillow==10.4.0
requests
scipy==1.11.4
moviepy==1.0.3
tensorflow-cpu==2.15.0
psycopg2-binary
bcrypt
nltk
SQLAlchemy
>>>>>>> c4839494da2a5a9df69db6655d4e7b1395253724
